fix: concurrency, format specifier, gmtime_r thread-safety in plugin
